Fix intellij warnings

+++ holmes-rome-modules/src/main/java/com/sun/syndication/feed/module/mediarss/types/MediaGroup.java
@@ -1,4 +1,21 @@
 /*
+ * Copyright (C) 2012-2013  Cedric Cheneau
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/*
  * MediaGroup.java
  *
  * Created on April 19, 2006, 12:34 AM
@@ -23,15 +40,16 @@
  */
 package com.sun.syndication.feed.module.mediarss.types;
 
-import java.io.Serializable;
-
 import com.sun.syndication.feed.impl.EqualsBean;
 import com.sun.syndication.feed.impl.ToStringBean;
+
+import java.io.Serializable;
 
 /**
  * <strong>&lt;media:group&gt;</strong></p>
  * <p>&lt;media:group&gt; is a sub-element of &lt;item&gt;. It allows grouping of &lt;media:content&gt; elements that are effectively the same content, yet different representations.&nbsp;For instance: the same song recorded in both the WAV and MP3 format.
  * It's an optional element that must only be used for this purpose.</p>
+ *
  * @author cooper
  */
 public final class MediaGroup implements Cloneable, Serializable {
@@ -43,6 +61,7 @@ public final class MediaGroup implements Cloneable, Serializable {
 
     /**
      * Creates a new instance of MediaGroup
+     *
      * @param contents Contents of the group.
      */
     public MediaGroup(final MediaContent[] contents) {
@@ -51,7 +70,8 @@ public final class MediaGroup implements Cloneable, Serializable {
 
     /**
      * Creates a new instance of MediaGroup
-     * @param contents contents of the group
+     *
+     * @param contents            contents of the group
      * @param defaultContentIndex index of the default content value.
      */
     public MediaGroup(final MediaContent[] contents, final Integer defaultContentIndex) {
@@ -61,9 +81,10 @@ public final class MediaGroup implements Cloneable, Serializable {
 
     /**
      * Creates a new instance of MediaGroup
-     * @param contents contents of the group
+     *
+     * @param contents            contents of the group
      * @param defaultContentIndex index of the default content item.
-     * @param metadata metadata for the group.
+     * @param metadata            metadata for the group.
      */
     public MediaGroup(final MediaContent[] contents, final Integer defaultContentIndex, final Metadata metadata) {
         this.setContents(contents);
@@ -73,6 +94,7 @@ public final class MediaGroup implements Cloneable, Serializable {
 
     /**
      * MediaContents for the group
+     *
      * @param contents MediaContents for the group
      */
     public void setContents(final MediaContent[] contents) {
@@ -81,6 +103,7 @@ public final class MediaGroup implements Cloneable, Serializable {
 
     /**
      * MediaContents for the group
+     *
      * @return MediaContents for the group
      */
     public MediaContent[] getContents() {
@@ -89,11 +112,12 @@ public final class MediaGroup implements Cloneable, Serializable {
 
     /**
      * Default content index MediaContent.
+     *
      * @param defaultContentIndex Default content index MediaContent.
      */
     public void setDefaultContentIndex(final Integer defaultContentIndex) {
         for (int i = 0; i < getContents().length; i++) {
-            if (i == defaultContentIndex.intValue()) {
+            if (i == defaultContentIndex) {
                 getContents()[i].setDefaultContent(true);
             } else {
                 getContents()[i].setDefaultContent(false);
@@ -105,6 +129,7 @@ public final class MediaGroup implements Cloneable, Serializable {
 
     /**
      * Default content index MediaContent.
+     *
      * @return Default content index MediaContent.
      */
     public Integer getDefaultContentIndex() {
@@ -113,6 +138,7 @@ public final class MediaGroup implements Cloneable, Serializable {
 
     /**
      * Metadata for the group
+     *
      * @param metadata Metadata for the group
      */
     public void setMetadata(final Metadata metadata) {
@@ -121,6 +147,7 @@ public final class MediaGroup implements Cloneable, Serializable {
 
     /**
      * Metadata for the group
+     *
      * @return Metadata for the group
      */
     public Metadata getMetadata() {
